backend/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "resume-vacancy-api"
version = "1.0.0"
description = "Django REST API for Resume/Vacancy Management"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "Resume API Team"}]
license = {text = "MIT"}

dependencies = [
    "Django>=4.2,<5.0",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.3.0",
    "django-environ>=0.11.2",
    "djangorestframework-simplejwt>=5.3.0",
    "drf-spectacular>=0.27.0",
    "drf-spectacular[sidecar]>=0.27.0",
    "Pillow>=10.1.0",
    "psycopg[binary]>=3.3.1",
    "gunicorn>=21.2.0",
    "whitenoise>=6.6.0",
    "python-decouple>=3.8",
    "structlog>=23.2.0",
    "sentry-sdk>=1.39.1",
    "django-storages[s3]>=1.14.2",
    "boto3>=1.34.0",
    "celery>=5.3.4",
    "redis>=5.0.1",
    "django-ratelimit>=4.1.0",
    "django-filter>=23.5",
    "django-extensions==3.2.3",
    "PyMuPDF>=1.24.14",
    "python-docx>=1.1.2",
    "reportlab>=4.2.5",
    "reportlab==4.2.5",
    "openai>=1.97.0",
    "Telethon>=1.36.0",
    "pgvector>=0.3.6",
    "psutil>=5.9.0",
]


[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-django>=4.7.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.11",
    "isort>=5.13.2",
    "flake8>=6.1.0",
    "faker>=22.6.0",
    "factory-boy>=3.3.0",
    "responses>=0.24.1",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.dev"
pythonpath = "."
testpaths = ["tests", "apps"]
addopts = "--strict-markers --cov=apps --cov-report=html --cov-report=term-missing"
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_mode = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["*/migrations/*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "RUF"]
ignore = [
    "E501",     # line length handled by formatter
    "RUF001",   # ambiguous unicode — expected in ru/uz strings
    "RUF002",   # ambiguous unicode in docstrings — emojis in parser docs
    "RUF012",   # mutable class attrs — false positive on Django models/serializers/migrations
]

[tool.ruff.lint.per-file-ignores]
"*/migrations/*" = ["E", "F", "W", "I", "N", "UP", "RUF"]

[tool.coverage.run]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "manage.py",
    "config/wsgi.py",
    "config/asgi.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[dependency-groups]
dev = [
    "ruff>=0.15.7",
    "pytest>=7.4.3",
    "pytest-django>=4.7.0",
    "pytest-cov>=4.1.0",
    "faker>=22.6.0",
    "factory-boy>=3.3.0",
    "responses>=0.24.1",
]
